Added standatd python packaging structure and information
* added setup.py
* moved sources to sleepymongoose directory
* linked httpd.py (for backwards compatibility)
* added requirements.txt
